Rename internal gbm package
It conflicts with pkg/gbm and the internal package is really just a util
