+++ cli/internal/utils/utils.go
@@ -1,10 +1,26 @@
-package release
+package utils
 
 import (
 	"fmt"
+	"regexp"
+	"time"
 
 	"github.com/wordpress-mobile/gbm-cli/internal/repo"
 )
+
+func IsScheduledRelease(version string) bool {
+	re := regexp.MustCompile(`^v*(\d+)\.(\d+)\.0$`)
+	return re.MatchString(version)
+}
+
+func NextReleaseDate() string {
+	weekday := time.Now().Weekday()
+	daysUntilThursday := 4 - weekday
+
+	nextThursday := time.Now().AddDate(0, 0, int(daysUntilThursday))
+
+	return nextThursday.Format("Monday 01, 2006")
+}
 
 func GetGbmReleasePr(version string) (repo.PullRequest, error) {
 	filter := repo.BuildRepoFilter("gutenberg-mobile", "is:pr", fmt.Sprintf("%s in:title", version))
